+++ lib/src/ats/src/comand_execute_history.dart
@@ -4,56 +4,56 @@ part of '../ats.dart';
 abstract class AtsCommandExecutionHistory with _$AtsCommandExecutionHistory {
   /// CommandExecutionHistory entity definition.
   factory AtsCommandExecutionHistory({
-    /// ID of the command execution history. This ID is unique.
+    /// [id]: ID of the command execution history. This ID is unique.
     String? id,
 
-    /// UUID of the command
+    /// [commandId]: UUID of the command
     String? commandId,
 
-    /// Asset from which the command was sent
-    String? fromAsset,
+    /// [fromAsset]: Asset from which the command was sent.
+    Asset? fromAsset,
 
-    /// Asset to which the command was sent
-    String? toAsset,
+    /// [toAsset]: Asset to which the command was sent.
+    Asset? toAsset,
 
-    /// Status of the command execution
+    /// [status]: Status of the command execution
     @JsonKey(unknownEnumValue: AtsCommandExecutionStatus.unknown)
     @Default(AtsCommandExecutionStatus.unknown)
     AtsCommandExecutionStatus status,
 
-    /// Application from which the command was sent
+    /// [fromApp]: Application from which the command was sent
     @JsonKey(unknownEnumValue: FromApp.unknown) @Default(FromApp.unknown) FromApp fromApp,
 
-    /// Father asset related to the command execution
-    String? fatherAsset,
+    /// [fatherAsset]: Father asset related to the command execution.
+    Asset? fatherAsset,
 
-    /// Error response if the command failed
+    /// [errorResponse] Error response if the command failed
     String? errorResponse,
 
-    /// User who generated the command execution
-    String? generatedBy,
+    /// [generatedBy]: User who generated the command execution.
+    User? generatedBy,
 
-    /// Queue ID associated with the command execution
+    /// [queueId]: Queue ID associated with the command execution
     String? queueId,
 
-    /// Mileage of the to_asset at the time of command execution
+    /// [toAssetMileage]: Mileage of the to_asset at the time of command execution
     double? toAssetMileage,
 
-    /// Timestamp when the command execution history was created (Unix)
+    /// [createdAt]: Timestamp when the command execution history was created (Unix)
     @TimestampConverter() DateTime? createdAt,
 
-    /// Timestamp when the command execution history was last updated (Unix)
+    /// [updatedAt]: Timestamp when the command execution history was last updated (Unix).
     @TimestampConverter() DateTime? updatedAt,
 
-    /// Payload of the command execution
+    /// [payload]: Payload of the command execution
     String? payload,
 
-    /// Type of the command executed
+    /// [commandType]: Type of the command executed
     @JsonKey(unknownEnumValue: AtsCommandExecutionType.unknown)
     @Default(AtsCommandExecutionType.unknown)
     AtsCommandExecutionType commandType,
 
-    /// Communication method used for the command execution
+    /// [communicationMethod]: Communication method used for the command execution
     @JsonKey(unknownEnumValue: AtsCommunicationMethod.unknown)
     @Default(AtsCommunicationMethod.unknown)
     AtsCommunicationMethod communicationMethod,